Simplify embeddings config/flow
