+++ graphrag/config/models/text_embedding_config.py
@@ -6,7 +6,6 @@
 from pydantic import BaseModel, Field
 
 from graphrag.config.defaults import graphrag_config_defaults
-from graphrag.config.enums import TextEmbeddingTarget
 from graphrag.config.models.language_model_config import LanguageModelConfig
 
 
@@ -28,10 +27,6 @@ class TextEmbeddingConfig(BaseModel):
     batch_max_tokens: int = Field(
         description="The batch max tokens to use.",
         default=graphrag_config_defaults.embed_text.batch_max_tokens,
-    )
-    target: TextEmbeddingTarget = Field(
-        description="The target to use. 'all', 'required', 'selected', or 'none'.",
-        default=graphrag_config_defaults.embed_text.target,
     )
     names: list[str] = Field(
         description="The specific embeddings to perform.",
